deploy/sync-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-firestore==2.14.0
google-generativeai==0.3.2
httpx==0.26.0
python-dotenv==1.0.0
